Plugins package refactoring: split client and server plugins
