+++ src/main/kotlin/com/example/plugins/client/Configuration.kt
@@ -1,4 +1,4 @@
-package com.example.plugins
+package com.example.plugins.client
 
 import io.ktor.client.*
 import io.ktor.client.engine.cio.*
