+++ src/main/kotlin/com/example/plugins/server/Logging.kt
@@ -1,4 +1,4 @@
-package com.example.plugins
+package com.example.plugins.server
 
 import io.ktor.server.application.*
 import io.ktor.server.plugins.callloging.*
